Avoid using tmpnam() for security reasons. run_process_capture_output() now checks whether the child process' exit code was zero.

+++ Scripts/PerlModules/ProcessUtils.pm
@@ -7,6 +7,7 @@ use strict;
 use warnings;
 
 use File::Temp;
+use Fcntl qw/F_SETFD F_GETFD FD_CLOEXEC/;
 
 use FileUtils;
 use StringUtils;
@@ -73,15 +74,54 @@ sub reason_died_from_wait_code ( $ )
 }
 
 
+sub clear_cloexec_flag ( $ )
+{
+  my $fh = shift;
+
+  my $oldFlags = fcntl( $fh, F_GETFD, 0 );
+  my $newFlags = $oldFlags & ~FD_CLOEXEC;
+
+  fcntl( $fh, F_SETFD, $newFlags ) or
+    die "Cannot clear close-on-exec flag (FD_CLOEXEC) on file descriptor: $!\n";
+}
+
+
 sub run_process_capture_output ( $ $ $ )
 {
   my $cmd_line = shift;
   my $stdout_capture = shift;  # ref to array
   my $stderr_capture = shift;  # ref to array
 
-  # TODO: for security reasons, try to avoid tmpnam().
-  my $child_stdout_file_name = tmpnam();
-  my $child_stderr_file_name = tmpnam();
+  use constant USE_TMPNAM => 0;  # tmpnam() has known security and timing issues.
+
+  my $child_stdout_file_name;
+  my $child_stderr_file_name;
+
+  my $stdoutFile;
+  my $stderrFile;
+
+  if ( USE_TMPNAM )
+  {
+    $child_stdout_file_name = tmpnam();
+    $child_stderr_file_name = tmpnam();
+  }
+  else
+  {
+    $stdoutFile = File::Temp->new( UNLINK => 1 );
+    $stderrFile = File::Temp->new( UNLINK => 1 );
+
+    clear_cloexec_flag( $stdoutFile );
+    clear_cloexec_flag( $stderrFile );
+
+    # For standard applications, the documentation for File::Temp suggests using a filename like /dev/fd/123,
+    # but Bash recognises those filenames and handles them differently, so we cannot do that here.
+    # This is an excerpt from the bash documentation:
+    #    Bash handles several filenames specially when they are used in redirections, as described in the following table:
+    #      /dev/fd/fd    If fd is a valid integer, file descriptor fd is duplicated.
+
+    $child_stdout_file_name = "&" . fileno( $stdoutFile );
+    $child_stderr_file_name = "&" . fileno( $stderrFile );
+  }
 
   my $cmd_line_to_use = $cmd_line . " >$child_stdout_file_name 2>$child_stderr_file_name"; 
 
@@ -89,24 +129,41 @@ sub run_process_capture_output ( $ $ $ )
 
   eval
   {
-    @$stdout_capture = FileUtils::read_text_file( $child_stdout_file_name );
-    @$stderr_capture = FileUtils::read_text_file( $child_stderr_file_name );
+    if ( USE_TMPNAM )
+    {
+      @$stdout_capture = FileUtils::read_text_file( $child_stdout_file_name );
+      @$stderr_capture = FileUtils::read_text_file( $child_stderr_file_name );
+    }
+    else
+    {
+      $stdoutFile->seek( 0, SEEK_SET ) or die "Cannot seek to the beginning of the file: $!\n";
+      $stderrFile->seek( 0, SEEK_SET ) or die "Cannot seek to the beginning of the file: $!\n";
+
+      @$stdout_capture = readline( $stdoutFile );
+      @$stderr_capture = readline( $stderrFile );
+    }
   };
 
   if ( $@ )
   {
     die qq<Error reading captured output from child process, > .
-        qq<the previous system() call probably failed for command line "$cmd_line", > .
+        qq<the previous system() call probably failed for command line "$cmd_line_to_use", > .
         qq<the error is: $@>;
   }
 
-  unlink( $child_stdout_file_name ) # Hopes that unlink() reports errors like open(), etc.
-    or die "Cannot delete file \"$child_stdout_file_name\": $!"; 
+  if ( USE_TMPNAM )
+  {
+    unlink( $child_stdout_file_name ) # Hopes that unlink() reports errors like open(), etc.
+        or die "Cannot delete file \"$child_stdout_file_name\": $!"; 
 
-  unlink( $child_stderr_file_name ) # Hopes that unlink() reports errors like open(), etc.
-    or die "Cannot delete file \"$child_stderr_file_name\": $!"; 
+    unlink( $child_stderr_file_name ) # Hopes that unlink() reports errors like open(), etc.
+        or die "Cannot delete file \"$child_stderr_file_name\": $!"; 
+  }
 
-  return $exit_code;
+  if ( $exit_code != 0 )
+  {
+    die "Error running process, the exit code was $exit_code, the command line was: $cmd_line_to_use\n";
+  }
 }
